Move models to their own files
The problem I'm seeing when I replicate https://github.com/flippercloud/flipper/issues/942 is fixed by this.

The specs and tests still pass.

And when I run Flipper.adapter.get_all from rails runner or console I get the same result.

Hoping this covers all the scenarios.

ref https://github.com/flippercloud/flipper/pull/832
ref https://github.com/flippercloud/flipper/pull/935
ref https://github.com/flippercloud/flipper/pull/924
