+++ lib/flipper/adapters/active_record/model.rb
@@ -1,0 +1,9 @@
+module Flipper
+  module Adapters
+    class ActiveRecord
+      class Model < ::ActiveRecord::Base
+        self.abstract_class = true
+      end
+    end
+  end
+end
